\name{pREC-package}
\alias{pREC-package}
\alias{pREC}
\docType{package}
\title{Probabilistic Regional Envelope Curve
}
\description{
This package aims to compute the regional envelope curve of extremes regional values with an assigned return period (i.e. non exceedence probabilities). 
}
\details{
\tabular{ll}{
Package: \tab pREC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-15\cr
License: \tab GPLv3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Attilio Castellarin, Bjorn Guse and Alessio Pugliese

Maintainer: Alessi Pugliese <alessio.pugliese3@unibo.it>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
